Se corrigen errores minimos. Se modulariza correctamente el programa. Y se inicia y termina los método de sobremuestreo
