+++ Codigo_Fuente/Algorithm_Resize_Img/Algorithm_Resize_Img.pro
@@ -5,6 +5,7 @@ CONFIG -= app_bundle
 
 SOURCES += \
         imageresize.cpp \
+        img_rsz_functs.cpp \
         main.cpp
 
 HEADERS += \
