+++ Codigo_Fuente/Algorithm_Resize_Img/BD/ImgResize.txt
@@ -1,57 +1,57 @@
 uint8_t RedMatrix [256] = { 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 
-205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 205, 
-254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 200, 248, 254, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 237, 0, 212, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 203, 197, 54, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 214, 200, 211, 58, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 194, 220, 54, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 80, 218, 254, 148, 254, 254, 254, 254, 254, 254, 254, 254, 
-249, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 
-205, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204, 204
+0, 0, 2, 21, 48, 48, 48, 48, 17, 17, 4, 4, 0, 0, 2, 2, 
+0, 0, 2, 21, 48, 48, 48, 48, 17, 17, 4, 4, 0, 0, 2, 2, 
+9, 13, 20, 59, 147, 147, 149, 149, 63, 63, 20, 20, 13, 13, 14, 14, 
+9, 13, 20, 59, 147, 147, 149, 149, 63, 63, 20, 20, 13, 13, 14, 14, 
+33, 40, 135, 175, 33, 33, 0, 0, 134, 134, 147, 147, 44, 44, 32, 32, 
+33, 40, 135, 175, 33, 33, 0, 0, 134, 134, 147, 147, 44, 44, 32, 32, 
+28, 155, 208, 54, 25, 25, 80, 80, 108, 108, 203, 203, 154, 154, 27, 27, 
+28, 155, 208, 54, 25, 25, 80, 80, 108, 108, 203, 203, 154, 154, 27, 27, 
+10, 24, 125, 136, 13, 13, 12, 12, 143, 143, 129, 129, 25, 25, 16, 16, 
+10, 24, 125, 136, 13, 13, 12, 12, 143, 143, 129, 129, 25, 25, 16, 16, 
+5, 8, 27, 69, 163, 163, 159, 159, 68, 68, 29, 29, 12, 12, 4, 4, 
+5, 8, 27, 69, 163, 163, 159, 159, 68, 68, 29, 29, 12, 12, 4, 4, 
+5, 8, 27, 69, 163, 163, 159, 159, 68, 68, 29, 29, 12, 12, 4, 4, 
+0, 3, 11, 18, 27, 27, 24, 24, 21, 21, 13, 13, 4, 4, 2, 2, 
+0, 3, 11, 18, 27, 27, 24, 24, 21, 21, 13, 13, 4, 4, 2, 2, 
+0, 3, 11, 18, 27, 27, 24, 24, 21, 21, 13, 13, 4, 4, 2, 2
 }; 
 
 uint8_t GreenMatrix [256] = { 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 0, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 220, 2, 6, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 172, 197, 54, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 0, 1, 211, 56, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 254, 175, 6, 54, 254, 254, 254, 254, 254, 254, 254, 254, 
-254, 254, 254, 254, 79, 218, 254, 118, 254, 254, 254, 254, 254, 254, 254, 254, 
-249, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 254, 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
+157, 157, 153, 148, 137, 137, 138, 138, 149, 149, 155, 155, 152, 152, 153, 153, 
+157, 157, 153, 148, 137, 137, 138, 138, 149, 149, 155, 155, 152, 152, 153, 153, 
+154, 150, 141, 157, 217, 217, 222, 222, 161, 161, 141, 141, 144, 144, 153, 153, 
+154, 150, 141, 157, 217, 217, 222, 222, 161, 161, 141, 141, 144, 144, 153, 153, 
+146, 135, 195, 215, 70, 70, 27, 27, 170, 170, 204, 204, 139, 139, 147, 147, 
+146, 135, 195, 215, 70, 70, 27, 27, 170, 170, 204, 204, 139, 139, 147, 147, 
+138, 244, 254, 78, 43, 43, 98, 98, 127, 127, 245, 245, 244, 244, 140, 140, 
+138, 244, 254, 78, 43, 43, 98, 98, 127, 127, 245, 245, 244, 244, 140, 140, 
+150, 144, 205, 181, 29, 29, 29, 29, 188, 188, 208, 208, 140, 140, 149, 149, 
+150, 144, 205, 181, 29, 29, 29, 29, 188, 188, 208, 208, 140, 140, 149, 149, 
+160, 149, 139, 151, 213, 213, 210, 210, 151, 151, 140, 140, 147, 147, 152, 152, 
+160, 149, 139, 151, 213, 213, 210, 210, 151, 151, 140, 140, 147, 147, 152, 152, 
+160, 149, 139, 151, 213, 213, 210, 210, 151, 151, 140, 140, 147, 147, 152, 152, 
+154, 158, 151, 147, 145, 145, 143, 143, 150, 150, 153, 153, 154, 154, 157, 157, 
+154, 158, 151, 147, 145, 145, 143, 143, 150, 150, 153, 153, 154, 154, 157, 157, 
+154, 158, 151, 147, 145, 145, 143, 143, 150, 150, 153, 153, 154, 154, 157, 157
 }; 
 
 uint8_t BlueMatrix [256] = { 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
-51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 51, 
-51, 51, 51, 51, 51, 0, 66, 51, 51, 51, 51, 51, 51, 51, 51, 51, 
-51, 51, 51, 51, 50, 10, 0, 8, 51, 51, 51, 51, 51, 51, 51, 51, 
-51, 51, 51, 51, 51, 0, 197, 52, 51, 51, 51, 51, 51, 51, 51, 51, 
-51, 51, 51, 51, 10, 0, 213, 61, 51, 51, 51, 51, 51, 51, 51, 51, 
-51, 51, 51, 51, 51, 0, 16, 52, 51, 51, 51, 51, 51, 51, 51, 51, 
-51, 51, 51, 51, 15, 220, 59, 0, 51, 51, 51, 51, 51, 51, 51, 51, 
-53, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 56, 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
-2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
+58, 60, 60, 43, 7, 7, 6, 6, 37, 37, 60, 60, 67, 67, 74, 74, 
+58, 60, 60, 43, 7, 7, 6, 6, 37, 37, 60, 60, 67, 67, 74, 74, 
+51, 46, 38, 58, 128, 128, 130, 130, 58, 58, 35, 35, 48, 48, 60, 60, 
+51, 46, 38, 58, 128, 128, 130, 130, 58, 58, 35, 35, 48, 48, 60, 60, 
+32, 19, 71, 127, 63, 63, 20, 20, 80, 80, 75, 75, 19, 19, 28, 28, 
+32, 19, 71, 127, 63, 63, 20, 20, 80, 80, 75, 75, 19, 19, 28, 28, 
+25, 124, 125, 0, 57, 57, 110, 110, 37, 37, 111, 111, 120, 120, 22, 22, 
+25, 124, 125, 0, 57, 57, 110, 110, 37, 37, 111, 111, 120, 120, 22, 22, 
+45, 34, 84, 86, 2, 2, 0, 0, 85, 85, 81, 81, 37, 37, 58, 58, 
+45, 34, 84, 86, 2, 2, 0, 0, 85, 85, 81, 81, 37, 37, 58, 58, 
+59, 45, 27, 53, 152, 152, 144, 144, 43, 43, 22, 22, 52, 52, 68, 68, 
+59, 45, 27, 53, 152, 152, 144, 144, 43, 43, 22, 22, 52, 52, 68, 68, 
+59, 45, 27, 53, 152, 152, 144, 144, 43, 43, 22, 22, 52, 52, 68, 68, 
+51, 57, 54, 54, 59, 59, 53, 53, 47, 47, 46, 46, 55, 55, 63, 63, 
+51, 57, 54, 54, 59, 59, 53, 53, 47, 47, 46, 46, 55, 55, 63, 63, 
+51, 57, 54, 54, 59, 59, 53, 53, 47, 47, 46, 46, 55, 55, 63, 63
 };
