+++ Codigo_Fuente/Algorithm_Resize_Img/imageresize.cpp
@@ -1,66 +1,9 @@
 #include "imageresize.h"
 
-ImageResize::ImageResize(){}
-
-ImageResize::ImageResize(int _wLeds, int _hLeds, string _fileNameImg, string _fileNameTxt)
-{
-    // [ Inicializacion atributos básicos ]
-
-    mWidthL = _wLeds;
-    mHeightL = _hLeds;
-    mFileNameImg = _fileNameImg;
-    mFileNameTxt = _fileNameTxt;
-
-    IMG = new QImage(mFileNameImg.c_str());
-
-    // [ Inicializacion dinamica de matrices que contendrán los datos de la img redimensionada ]
-
-    mImgRed = new int *[mHeightL];
-    mImgGreen = new int *[mHeightL];
-    mImgBlue = new int *[mHeightL];
-    for (int i = 0; i < mHeightL; i++) {
-        mImgRed[i] = new int [mWidthL];
-        mImgGreen[i] = new int [mWidthL];
-        mImgBlue[i] = new int [mWidthL];
-    }
-}
-
-void ImageResize::categorizeAlgorithm()
-{
-    int width = IMG->width();
-    int height = IMG->height();
-
-//    float categW = width / mWidthL;
-//    float categH = height / mHeightL;
-
-    if ((width >= mWidthL) && (height >= mHeightL)) {
-        // En este caso se realiza submuestreo
-
-        if ((width % mWidthL == 0) && (height % mHeightL == 0)) {
-            // En este caso no se consideran pixeles sobrantes
-            subDivisionExact();
-        }else {
-            // Sí se consideran pixeles sobrantes
-            subNoDivisionExact();
-        }
-    }else if ((width <= mWidthL) && (height <= mWidthL)) {
-        // En este se hace sobremuestreo
-
-        if ((mWidthL % width == 0) && (mHeightL % height == 0)) {
-            // En este caso no se consideran pixeles faltantes
-        }else {
-            // Sí se consideran pixeles faltantes
-        }
-    }else if ((width >= mWidthL) && (height <= mHeightL)) {
-        // En este caso se hace submuestreo sobre el ancho,
-        // y sobremuestreo sobre el alto
-    }else if((width <= mWidthL) && (height >= mHeightL)) {
-        // En este caso de hace sobremuestreo sobre el ancho,
-        // y submuestreo sobre el alto
-    }
+// ImageResize::ImageResize(){}
 
 
-}
+/// [ METODOS DE SUBMUESTREO ]
 
 void ImageResize::subDivisionExact()
 {
@@ -89,10 +32,11 @@ void ImageResize::subNoDivisionExact()
     // posiblemente importante. Esta cantidad de pixeles que restan se obtienen con las variables:
     // pixResColum, pixResRow.
     // Entonces la forma de considerar estos pixeles restantes con que, cada vez que se coje un
-    // pixel dando su posicion, sumamos un a la coordenada x o y, dependiendo del caso.
+    // pixel dando su posicion, sumamos uno a la coordenada x o y, dependiendo del caso.
 
     int pixelsColum = width/mWidthL;
     int pixelsRow = height/mHeightL;
+
     // Indicadores de desde dónde empezar a considerar los pixeles restantes
     int pixResColum = width % mWidthL;
     int pixResRow = height % mHeightL;
@@ -136,91 +80,108 @@ void ImageResize::subNoDivisionExact()
 
 }
 
-
-void ImageResize::SaveImgTxt()
+/// [ METODOS DE SOBREMUESTREO
+void ImageResize::sobreDivisionExact()
 {
-    ofstream archImg(mFileNameTxt);
+    int width = IMG->width();
+    int height = IMG->height();
 
-    if (!archImg.is_open()) {
-        cout << "Error al abrir archivo" << endl;
-        exit(1);
-    }
+    int pixelsColum = mWidthL/width;
+    int pixelsRow = mHeightL/height;
 
-    // [ Se almacena el color Red ]
-    archImg << "uint8_t RedMatrix [256] = { \n";
-    for (int i = 0; i < mHeightL; i++) {
+    int red;
+    int green;
+    int blue;
 
-        for (int j = 0; j < mWidthL; j++) {
-            // Se considera el 255 por el error de tinkerkad
-            if (mImgRed[i][j] == 255) {
-                mImgRed[i][j] = 254;
-            }
-            // Considero el ultimo elemento, no debe llevar coma (,)
-            if ((i == mHeightL - 1) && (j == mWidthL - 1)) {
-                archImg << mImgRed[i][j];
-            }else {
-                archImg << mImgRed[i][j] << ", ";
-            }
-        }
-        archImg << "\n";
-    }
-    archImg << "}; \n\n";
-
-    // [ Se almacena el color Green ]
-    archImg << "uint8_t GreenMatrix [256] = { \n";
-    for (int i = 0; i < mHeightL; i++) {
-
-        for (int j = 0; j < mWidthL; j++) {
-            // Se considera el 255 por el error de tinkerkad
-            if (mImgGreen[i][j] == 255) {
-                mImgGreen[i][j] = 254;
-            }
-            // Considero el ultimo elemento, no debe llevar coma (,)
-            if ((i == mHeightL - 1) && (j == mWidthL - 1)) {
-                archImg << mImgGreen[i][j];
-            }else {
-                archImg << mImgGreen[i][j] << ", ";
+    for (int i = 0; i < height; i++) {
+        for (int j = 0; j < width; j++) {
+            // Se coge el color a replicar
+            red = IMG->pixelColor(j, i).red();
+            green = IMG->pixelColor(j, i).green();
+            blue = IMG->pixelColor(j, i).blue();
+            // En la matriz de leds, se pone repetitivamene tal color
+            for (int iCopy = 0; iCopy < pixelsRow; iCopy++) {
+                for (int jCopy = 0; jCopy < pixelsColum; jCopy++) {
+                    mImgRed[i*pixelsRow + iCopy][j*pixelsColum + jCopy] = red;
+                    mImgGreen[i*pixelsRow + iCopy][j*pixelsColum + jCopy] = green;
+                    mImgBlue[i*pixelsRow + iCopy][j*pixelsColum + jCopy] = blue;
+                }
             }
         }
-        archImg << "\n";
     }
-    archImg << "}; \n\n";
-
-    // [ Se almacena el color Blue ]
-    archImg << "uint8_t BlueMatrix [256] = { \n";
-    for (int i = 0; i < mHeightL; i++) {
-
-        for (int j = 0; j < mWidthL; j++) {
-            // Se considera el 255 por el error de tinkerkad
-            if (mImgBlue[i][j] == 255) {
-                mImgBlue[i][j] = 254;
-            }
-            // Considero el ultimo elemento, no debe llevar coma (,)
-            if ((i == mHeightL - 1) && (j == mWidthL - 1)) {
-                archImg << mImgBlue[i][j];
-            }else {
-                archImg << mImgBlue[i][j] << ", ";
-            }
-        }
-        archImg << "\n";
-    }
-    archImg << "}; \n\n";
-
-    archImg.close();
-
 }
 
-
-ImageResize::~ImageResize()
+void ImageResize::sobreNoDivisionExact()
 {
-    for (int i = 0; i < mHeightL; i++) {
-        delete [] mImgRed[i];
-        delete [] mImgGreen[i];
-        delete [] mImgBlue[i];
+    int width = IMG->width();
+    int height = IMG->height();
+
+    // Indican la cantidad de veces a replicar un pixel, ya sea de columna o fila
+    int pixelsColum = mWidthL/width;
+    int pixelsRow = mHeightL/height;
+
+    // Indican la cantidad de pixels faltantes para completar la matriz de leds
+    int pixResColum = mWidthL % width;
+    int pixResRow = mHeightL % height;
+
+    // Contadores auxuliares para el relleno de pixeles en filas y columnas
+    // Se inicializan en -1 porque en caso contrario indexan la matriz en posiciones inválidas
+    int contPixR = -1;
+    int contPixC = -1;
+
+    int red;
+    int green;
+    int blue;
+
+    for (int i = 0; i < height; i++) {
+        if ((height - pixResRow) <= i) {
+            contPixR++;
+        }
+
+        for (int j = 0; j < width; j++) {
+            red = IMG->pixelColor(j, i).red();
+            green = IMG->pixelColor(j, i).green();
+            blue = IMG->pixelColor(j, i).blue();
+
+            // Caso en que rellenan con pixeles en la filas(row) y las columnas(colum)
+            if ((height - pixResRow) <= i  && (width - pixResColum) <= j) {
+                contPixC++;
+                for (int iCopy = 0; iCopy < pixelsRow + 1; iCopy++) {
+                    for (int jCopy = 0; jCopy < pixelsColum + 1; jCopy++) {
+                        mImgRed[i*pixelsRow + iCopy + contPixR][j*pixelsColum + jCopy + contPixC] = red;
+                        mImgGreen[i*pixelsRow + iCopy + contPixR][j*pixelsColum + jCopy + contPixC] = green;
+                        mImgBlue[i*pixelsRow + iCopy + contPixR][j*pixelsColum + jCopy + contPixC] = blue;
+                    }
+                }
+            }// Caso en que no se rellenan con pixeles ni en las filas(row) ni en las columnas(colum)
+            else if ((height - pixResRow) > i && (width - pixResColum) > j) {
+                for (int iCopy = 0; iCopy < pixelsRow; iCopy++) {
+                    for (int jCopy = 0; jCopy < pixelsColum; jCopy++) {
+                        mImgRed[i*pixelsRow + iCopy][j*pixelsColum + jCopy] = red;
+                        mImgGreen[i*pixelsRow + iCopy][j*pixelsColum + jCopy] = green;
+                        mImgBlue[i*pixelsRow + iCopy][j*pixelsColum + jCopy] = blue;
+                    }
+                }
+            }// Caso en que se rellenan con pixeles solo en las filas(row)
+            else if ((height - pixResRow) <= i) {
+                for (int iCopy = 0; iCopy < pixelsRow + 1; iCopy++) {
+                    for (int jCopy = 0; jCopy < pixelsColum; jCopy++) {
+                        mImgRed[i*pixelsRow + iCopy + contPixR][j*pixelsColum + jCopy] = red;
+                        mImgGreen[i*pixelsRow + iCopy + contPixR][j*pixelsColum + jCopy] = green;
+                        mImgBlue[i*pixelsRow + iCopy + contPixR][j*pixelsColum + jCopy] = blue;
+                    }
+                }
+            }// Caso en que se rellenan con pixeles solo en las columnas(colum)
+            else {
+                contPixC++;
+                for (int iCopy = 0; iCopy < pixelsRow; iCopy++) {
+                    for (int jCopy = 0; jCopy < pixelsColum + 1; jCopy++) {
+                        mImgRed[i*pixelsRow + iCopy][j*pixelsColum + jCopy + contPixC] = red;
+                        mImgGreen[i*pixelsRow + iCopy][j*pixelsColum + jCopy + contPixC] = green;
+                        mImgBlue[i*pixelsRow + iCopy][j*pixelsColum + jCopy + contPixC] = blue;
+                    }
+                }
+            }
+        }contPixC = -1;
     }
-    delete [] mImgRed;
-    delete [] mImgGreen;
-    delete [] mImgBlue;
-
-
 }
