+++ Codigo_Fuente/Algorithm_Resize_Img/main.cpp
@@ -9,7 +9,7 @@ using namespace std;
 int main()
 {
 
-    string fileName = "../Algorithm_Resize_Img/BD/espana.jpg";
+    string fileName = "../Algorithm_Resize_Img/BD/brasil10x7.jpg";
     string fileTxt = "../Algorithm_Resize_Img/BD/ImgResize.txt";
 
     ImageResize imgR(16, 16, fileName, fileTxt);
